Código más limpio de la base de datos tanto para la pantalla comidas como la pantalla training
